Add files via upload
then this is the create plan page.... and apart from the intermitent error with the initial firing of it this was working fine but it creates a new plan for every customer! I have already created a plan within paypal that i want EVERY customer to link to the same paln so figured this approach was wrong, there is no docs that i could find relating to this at all on paypal so im kind of winging it from her but i decided to go with what i thought.... and i have actually got to the same stage by linking to my paln so first get token then link my plan details in the paypal button script and this all goes thru but same two issues... il put that script up next just wanted to show this bit working ok.